Add hover effects and responsive design
