+++ src/assets/stylesheets/abstracts/_mixins.scss
@@ -71,4 +71,19 @@
   padding: 5rem;
   padding-bottom: 0;
   transform: translateX(10rem);
+
+  @include respond(tablet-landscape) {
+    width: 85rem;
+  }
+
+  @include respond(tablet-portrait) {
+    padding: 2rem;
+    padding-bottom: 0;
+    width: 75rem;
+  }
+
+  @include respond(tablet-portrait) {
+    width: 86rem;
+    transform: translateX(0);
+  }
 }
