+++ src/assets/stylesheets/components/_DashboardNav.scss
@@ -1,4 +1,5 @@
 @import "../abstracts/variables";
+@import "../abstracts/mixins";
 
 #dashboard-nav {
   display: flex;
@@ -6,23 +7,38 @@
   height: 100vh;
   position: fixed;
   width: fit-content;
-  padding-right: 2rem;
   z-index: 98;
   left: 0;
   background-color: $color-dashboard-white;
   transition: all 0.3s ease;
   box-shadow: rgba(0, 0, 0, 0.08) 0 0.5rem 1rem;
 
+  nav {
+    height: 100%;
+  }
+
+  @include respond(tablet-portrait) {
+    display: none;
+  }
+
   .nav-list {
-    margin-top: 5rem;
+    margin-top: 3rem;
   }
 
   .nav-item {
     display: flex;
     flex-direction: row;
+    align-items: center;
     list-style: none;
-    margin-left: 2rem;
-    margin-top: 2.5rem;
+    padding: 1.75rem;
+    cursor: pointer;
+    transition: all 0.3s;
+    margin: 1rem 0rem;
+    margin-left: 1rem;
+
+    &:hover {
+      background-color: $color-primary-light;
+    }
 
     a {
       display: flex;
@@ -36,11 +52,17 @@
       }
 
       svg {
-        font-size: 2.5rem;
+        font-size: 2.2rem;
       }
+    }
+  }
 
-      p {
-        font-size: 2rem;
+  .nav-item-active {
+    background-color: $color-primary !important;
+    a {
+      p,
+      svg {
+        color: $color-white;
       }
     }
   }
