+++ src/assets/stylesheets/pages/_p_dashboard.scss
@@ -2,6 +2,6 @@
 @import "../abstracts/variables";
 
 #dashboard {
-  justify-content: center;
   display: flex;
+  justify-content: center;
 }
