+++ src/components/MobileMenuItem.tsx
@@ -1,7 +1,6 @@
 import "../assets/stylesheets/components/_MobileMenuItem.scss";
 import { Link } from "react-router-dom";
-import { useAuth, useAuthUpdate } from "../contexts/AuthContext";
-import { toast } from "react-toastify";
+import { useAuth } from "../contexts/AuthContext";
 
 interface MobileMenuItemProps {
   to: string;
@@ -10,29 +9,15 @@ interface MobileMenuItemProps {
 
 const MobileMenuItem: React.FC<MobileMenuItemProps> = ({ to, resource }) => {
   const authStatus = useAuth();
-  const setAuth = useAuthUpdate();
-
-  const handleLogout = () => {
-    localStorage.removeItem("token");
-    setAuth(false);
-    toast.success(`Successfully Logged Out!`);
-  };
 
   return (
     <div className="mobile-menu-item">
-      {authStatus ? (
+      {authStatus && (
         <>
           <Link to={to}>
             <li className="mobile-menu-item">{resource}</li>
           </Link>
-          <Link to="/" onClick={handleLogout}>
-            <li className="mobile-menu-item">Logout</li>
-          </Link>
         </>
-      ) : (
-        <Link to={`/login`}>
-          <li className="mobile-menu-item">Login</li>
-        </Link>
       )}
     </div>
   );
